app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
google-cloud-vision==3.7.2
google-cloud-firestore==2.16.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0